requirements.txt
openai
pandas
python-dotenv
sqlite3
streamlit